+++ src/framework/graph-visualiser/explorer/graph-instance-explorer.component.html
@@ -7,27 +7,27 @@
     </div>
   } @else {
     <div class="detail-panel explorer-detail">
-      @if (vm.isDataStale) {
+      @if (state.isDataStale) {
         <div class="stale-data-banner">
           <i class="fa-light fa-clock-rotate-left"></i>
           <span>Data was loaded in a different transaction</span>
-          <button mat-stroked-button (click)="vm.refresh()">
+          <button mat-stroked-button (click)="state.refresh()">
             <i class="fa-light fa-refresh"></i>
             Refresh
           </button>
         </div>
       }
       <div class="detail-content">
-        @if (vm.needsTransaction) {
+        @if (state.needsTransaction) {
           <div class="transaction-prompt">
             <i class="fa-light fa-lock"></i>
             <p>Open a transaction to explore your data</p>
-            <button mat-stroked-button (click)="vm.openTransactionAndLoad()">
+            <button mat-stroked-button (click)="state.openTransactionAndLoad()">
               <i class="fa-light fa-play"></i>
               Open read transaction
             </button>
           </div>
-        } @else if (vm.loading) {
+        } @else if (state.loading) {
           <div class="loading-container">
             <mat-spinner diameter="40"></mat-spinner>
           </div>
@@ -37,11 +37,11 @@
             <div class="type-row">
               <span class="type-info">
                 @if (type!.supertype) {
-                  <span class="type-label" [ngClass]="vm.typeKindLabel">{{ type!.label }}@if (type!.kind === 'attributeType') {<em class="value-type">, value {{ $any(type).valueType }}</em>}</span>
+                  <span class="type-label" [ngClass]="state.typeKindLabel">{{ type!.label }}@if (type!.kind === 'attributeType') {<em class="value-type">, value {{ $any(type).valueType }}</em>}</span>
                   <span class="type-muted">sub {{ type!.supertype.label }}</span>
                 } @else {
-                  <span class="type-muted">{{ vm.typeKindLabel }}</span>
-                  <span class="type-label" [ngClass]="vm.typeKindLabel">{{ type!.label }}@if (type!.kind === 'attributeType') {<em class="value-type">, value {{ $any(type).valueType }}</em>}</span>
+                  <span class="type-muted">{{ state.typeKindLabel }}</span>
+                  <span class="type-label" [ngClass]="state.typeKindLabel">{{ type!.label }}@if (type!.kind === 'attributeType') {<em class="value-type">, value {{ $any(type).valueType }}</em>}</span>
                 }
               </span>
               <span class="instance-iid">{{ instanceIID }}</span>
@@ -60,11 +60,11 @@
                 <h6>Links</h6>
               </button>
               @if (linksCollapsed) {
-              } @else if (vm.linksLoading) {
+              } @else if (state.linksLoading) {
                 <div class="relations-loading">
                   <mat-spinner diameter="24"></mat-spinner>
                 </div>
-              } @else if (vm.ownLinks.length === 0) {
+              } @else if (state.ownLinks.length === 0) {
                 <p class="text-muted">No links</p>
               } @else {
                 <div class="section-bulk-action">
@@ -73,7 +73,7 @@
                   </button>
                 </div>
                 <div class="links-list">
-                  @for (link of vm.ownLinks; track link.roleLabel + link.playerIID) {
+                  @for (link of state.ownLinks; track link.roleLabel + link.playerIID) {
                     <div class="link-card">
                       <div class="link-info">
                         <span class="role-label">{{ link.roleLabel }}</span>
@@ -116,7 +116,7 @@
                 <h6>Attributes</h6>
               </button>
               @if (attributesCollapsed) {
-              } @else if (vm.attributes.length === 0) {
+              } @else if (state.attributes.length === 0) {
                 <p class="text-muted">No attributes</p>
               } @else {
                 <div class="section-bulk-action section-bulk-action--tight">
@@ -126,7 +126,7 @@
                 </div>
                 <table class="attributes-table">
                   <tbody>
-                    @for (attr of vm.attributes; track attr.type) {
+                    @for (attr of state.attributes; track attr.type) {
                       <tr>
                         <td class="attr-label">{{ attr.type }}<em class="value-type">, value {{ attr.valueType }}</em>@if (isAttributeAdded(attr)) {<button
                             class="add-attr-icon-button added"
@@ -166,11 +166,11 @@
                 <h6>Relations</h6>
               </button>
               @if (relationsCollapsed) {
-              } @else if (vm.relationsLoading) {
+              } @else if (state.relationsLoading) {
                 <div class="relations-loading">
                   <mat-spinner diameter="24"></mat-spinner>
                 </div>
-              } @else if (vm.allRelations.length === 0) {
+              } @else if (state.allRelations.length === 0) {
                 <p class="text-muted">No relations</p>
               } @else {
                 <div class="section-bulk-action">
@@ -184,15 +184,15 @@
                       [class.active]="selectedRelationType === null"
                       (click)="selectRelationType(null)">
                     All
-                    <span class="chip-count">{{ vm.allRelations.length }}</span>
+                    <span class="chip-count">{{ state.allRelations.length }}</span>
                   </button>
-                  @for (relType of vm.relationTypes; track relType) {
+                  @for (relType of state.relationTypes; track relType) {
                     <button
                         class="filter-chip"
                         [class.active]="selectedRelationType === relType"
                         (click)="selectRelationType(relType)">
                       {{ relType }}
-                      <span class="chip-count">{{ vm.getRelationTypeCount(relType) }}</span>
+                      <span class="chip-count">{{ state.getRelationTypeCount(relType) }}</span>
                     </button>
                   }
                 </div>
@@ -305,15 +305,15 @@
                 <h6>Owners</h6>
               </button>
               @if (ownersCollapsed) {
-              } @else if (vm.ownersLoading) {
+              } @else if (state.ownersLoading) {
                 <div class="relations-loading">
                   <mat-spinner diameter="24"></mat-spinner>
                 </div>
-              } @else if (vm.owners.length === 0) {
+              } @else if (state.owners.length === 0) {
                 <p class="text-muted">No owners</p>
               } @else {
                 <div class="links-list">
-                  @for (owner of vm.owners; track owner.ownerIID) {
+                  @for (owner of state.owners; track owner.ownerIID) {
                     <div class="link-card">
                       <div class="link-info">
                         <div class="player-details">
