+++ src/framework/graph-visualiser/explorer/graph-instance-explorer.component.scss
