+++ src/framework/graph-visualiser/side-panel/graph-side-panel.component.html
@@ -53,7 +53,7 @@
                              [run]="run"
                              [visualiser]="visualiser"/>
     } @else {
-      <ts-graph-explorer
+      <ts-graph-instance-explorer
                         [type]="selectedType"
                         [instanceIID]="selectedInstanceIID"
                         [run]="run"
