+++ src/module/chat/chat-output/chat-output.component.html
@@ -138,7 +138,7 @@
                        [visualiser]="currentRun!.graph.visualiser ?? null"
                        [status]="currentRun!.graph.status"
                        [(maximised)]="graphMaximised"
-                       (canvasElRehomed)="onGraphCanvasRehomed($event)"
+                       (canvasElRebuilt)="onGraphCanvasRehomed($event)"
                        (statusAction)="onGraphStatusAction($event)">
       </ts-graph-canvas>
